Authentication done create,update,read and delete todos

+++ routers/todos.py
@@ -26,13 +26,19 @@ user_dependency = Annotated[dict,Depends(get_current_user)]
 
 @router.get("/")
 async def read_all(user: user_dependency,db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User not found")
+
     return db.query(Todos).filter(Todos.owner_id==user.get('user_id')).all()
 
 #Find a todo by ID:
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo_by_id(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo_by_id(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User not found")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('user_id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Item not found")
@@ -52,8 +58,13 @@ async def create_todo(user: user_dependency, db:db_dependency, todo_request: tod
     db.commit()
 
 @router.put("/todos/{todo_id}",status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency, todo_request: todo_Object, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: todo_Object, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User not found")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('user_id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Item not found")
     todo_model.title = todo_request.title
@@ -64,8 +75,12 @@ async def update_todo(db: db_dependency, todo_request: todo_Object, todo_id: int
 
 
 @router.delete("/todos/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User not found")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('user_id')).first()
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Item not found")
     db.delete(todo_model)
